Review Changes
IcmpEchoReply.java - Refactored variables to camelcase.

PingHandler.java - Removed the Ping class and created its own file in Ping.java

test.java - Refactored to PingTest.java

PingUtil.java, PingTest - Refactored datalen to dataLength
