+++ runelite-client/src/test/java/net/runelite/client/plugins/worldhopper/ping/PingTest.java
@@ -6,7 +6,7 @@ import java.net.InetAddress;
 import java.net.UnknownHostException;
 import org.junit.Test;
 
-public class test
+public class PingTest
 {
     @Test
     public void test() throws UnknownHostException
@@ -17,14 +17,14 @@ public class test
         InetAddress inetAddress = InetAddress.getByName("104.131.36.154");
         byte[] address = inetAddress.getAddress();
         String dataStr = "RuneLitePing";
-        int datalen = dataStr.length()+1;
-        Pointer data = new Memory(datalen);
+        int dataLength = dataStr.length()+1;
+        Pointer data = new Memory(dataLength);
         data.setString(0L, dataStr);
-        IcmpEchoReply icmpEchoReply = new IcmpEchoReply(new Memory(IcmpEchoReply.SIZE + datalen));
+        IcmpEchoReply icmpEchoReply = new IcmpEchoReply(new Memory(IcmpEchoReply.SIZE + dataLength));
         int packed = (address[0] & 0xff) | ((address[1] & 0xff)<<8) | ((address[2] & 0xff)<<16) | ((address[3] & 0xff)<<24);
-        int ret = ipHlpAPI.IcmpSendEcho(ptr, packed, data, (short)(datalen),0L, icmpEchoReply, IcmpEchoReply.SIZE + datalen, 1000);
+        int ret = ipHlpAPI.IcmpSendEcho(ptr, packed, data, (short)(dataLength),0L, icmpEchoReply, IcmpEchoReply.SIZE + dataLength, 1000);
         System.out.println("RET " + ret);
-        System.out.println("RTT " + icmpEchoReply.RoundTripTime.longValue());
+        System.out.println("RTT " + icmpEchoReply.roundTripTime.longValue());
         ipHlpAPI.IcmpCloseHandle(ptr);
     }
 }
